# handler/response.go
package handler

import (
	"fmt"
	"net/http"

	"github.com/Somaycon/atelie-api/schemas"
	"github.com/gin-gonic/gin"
)

func sendError(ctx *gin.Context, status int, message string) {
	ctx.Header("Content-type", "aplication/json")
	ctx.JSON(status, gin.H{
		"message" :message,
		"status": status,
	})
}

func sendSucess(ctx *gin.Context, op string, data interface{}) {
	ctx.Header("Content-type", "aplication/json")
	ctx.JSON(http.StatusOK, gin.H{
		"message": fmt.Sprintf("operation %s sucess", op),
		"data": data,
	})
}

type ErrorResponse struct {
	Message string `json:"message"`
	ErrorCode string `json:"errorCode"`
}

type CreateProductResponse struct {
	Message string `json:"message"`
	Data  schemas.ProductResponse `json:"data"`
}

type DeleteProductResponse struct {
	Message string `json:"message"`
	Data  schemas.ProductResponse `json:"data"`
}
type UpdateProductResponse struct {
	Message string `json:"message"`
	Data  schemas.ProductResponse `json:"data"`
}

type GetProductByIdResponse struct {
	Message string `json:"message"`
	Data  schemas.ProductResponse `json:"data"`
}

type GetAllProductsResponse struct {
	Message string `json:"message"`
	Data []schemas.ProductResponse `json:"data"`
}